moved compare_irf_cuts function to cuts.py and change it such that it returns false rather than raising exception when cuts differ
